refactor!(zaino-proto): make the utils expression-oriented, parse-don't-validate
prune_compact_block rebuilds each transaction through the pure
prune_compact_tx (per-field included_or_empty expressions) and fuses
the emptiness filter into the same iterator pipeline;
compact_block_to_nullifiers takes the same shape, with nullifiers_only
replacing the in-place mem::take dance. new_from_pool_types collects
its validation through Result's FromIterator, and provided_height
drops its Ok(...?) wrap.

ValidatedBlockRangeRequest now stores the u32s its endpoints parsed
to, making the range guarantee a type fact instead of a documented
promise; the consumer's `as u32` casts and their safety comments fall
away.

One deliberate semantics tightening, now that the filter is a set: a
request naming the same pool twice is rejected with the new
PoolTypeError::DuplicatePoolType instead of silently collapsing.

+++ packages/zaino-state/src/indexer/node_backed_indexer.rs
@@ -1256,9 +1256,8 @@ impl<Source: BlockchainSource> LightWalletIndexer for NodeBackedIndexerServiceSu
 
         let pool_type_filter = validated_request.pool_type_filter().clone();
 
-        // Note conversion here is safe due to the use of [`ValidatedBlockRangeRequest::new_from_block_range`]
-        let start = validated_request.start() as u32;
-        let end = validated_request.end() as u32;
+        let start = validated_request.start();
+        let end = validated_request.end();
 
         let service_clone = self.clone();
         let service_timeout = self.config.service.timeout;
@@ -1383,9 +1382,8 @@ impl<Source: BlockchainSource> LightWalletIndexer for NodeBackedIndexerServiceSu
 
         let pool_type_filter = validated_request.pool_type_filter().clone();
 
-        // Note conversion here is safe due to the use of [`ValidatedBlockRangeRequest::new_from_block_range`]
-        let start = validated_request.start() as u32;
-        let end = validated_request.end() as u32;
+        let start = validated_request.start();
+        let end = validated_request.end();
 
         let service_clone = self.clone();
         let service_timeout = self.config.service.timeout;
